Allowed types

Allowed matrix types

Davidson.AllowedTypes — Type

AllowedTypes = Union{Float32, Float64, ComplexF32, ComplexF64}

The allowed types T of the matrix whose eigenpairs are to be computed.

Allowed real types

Davidson.AllowedFloat — Type

AllowedFloat = Union{Float32, Float64}

The allowed real number types R that are compatible with the allowed matrix types T

Compatible types

The arrays of type R<:AllowedFloat supplied to the solver and solver! functions must be compatible with the arrays of type T<:AllowedTypes. For example, if T==ComplexF64, then it is required that R==Float64. The following table lists the compatible type pairs:

T	R

Float32	Float32

Float64	Float64

ComplexF32	Float32

ComplexF64	Float64