solver

This section details the use of the solver function, as well as its in-place version solver!. These compute a given number of the lowest-lying eigenpairs of a given matrix A in a 'matrix-free' manner. That is, only a function returning matrix-vector products is required as input, as opposed to the matrix itself.

Davidson.solver — Function

solver(f, diag, nroots, matdim; tol=1e-4, blocksize=nroots+5,
       maxvec=4*blocksize, niter=100, verbose=false)

Arguments

f::Function: In-place matrix-vector multiplication function

diag::Vector{T}: Diagonal of the matrix whose eigenpairs are sought, where T<:AllowedTypes

nroots::Int64: Number of eigenpairs to compute

matdim::Int64: Dimension of the matrix

Optional keyword arguments

tol::Float64: Residual norm convergence threshold

blocksize::Int64: Block size

maxvec::Int64: Maximum subspace dimension

niter::Int64: Maximum number of iterations

verbose::Bool: Verbose output flag. If true, then a summary is printed at the end of each iteration

Return values

The return value is of the form result = solver(…), where result is an EigenPairs object with the following fields:

vectors::Matrix{T}: Eigenvectors

values::Vector{R}: Eigenvalues

residuals::Vector{R}: Residual norms of the eigenvectors

converged::Vector{Bool}: Convergence flags (converged[i] == true if the i root converged)

Check for convergence

If verbose==false, no warning is printed if one or more roots fail to converge. The contents of result.converged should therefore be checked.

A note on the matrix-vector multiplication function

Let the matrix whose eigenpairs are sought be denoted by A. The in-place Function f recieves takes as arguments an AbstractMatrix{T} of vectors v and an AbstractMatrix{T} of matrix-vector products Av that is to be computed, where T<:AllowedTypes. The required form of this function is detailed in Matrix-vector multiplication function section.

Davidson.solver! — Function

solver!(vectors, values, f, diag, nroots, matdim, [Twork, Rwork];
        tol=1e-4, blocksize=nroots+5, maxvec=4*blocksize, niter=100,
        verbose=false, guess=false)

Arguments

vectors::Matrix{T}: Eigenvectors, where T<:AllowedTypes

values::Vector{R}: Eigenvalues, where R<:AllowedFloat

f::Function: In-place matrix-vector multiplication function

diag::Matrix{T}: Diagonal of the matrix whose eigenpairs are sought, where T<:AllowedTypes

nroots::Int64: Number of eigenpairs to compute

matdim::Int64: Dimension of the matrix

Optional arguments

The following two pre-allocated work arrays may be supplied:

Twork::Vector{T}: Type T<:AllowedTypes work array

Rwork::Vector{R}: Type R<:AllowedFloat work array

See Work arrays for the procedure for constructing these.

Optional keyword arguments

tol::Float64: Residual norm convergence threshold

blocksize::Int64: Block size

maxvec::Int64: Maximum subspace dimension

niter::Int64: Maximum number of iterations

verbose::Bool: Verbose output flag. If true, then a summary is printed at the end of each iteration

guess::Bool: If true, then on input, the vectors array is taken to contain the guess vectors

Return values

The return value is of the form result = solver!(…), where result is an ConvInfo object with the following fields:

residuals::Vector{R}: Residual norms of the eigenvectors

converged::Vector{Bool}: Convergence flags (converged[i] == true if the i root converged)

Check for convergence

If verbose==false, no warning is printed if one or more roots fail to converge. The contents of result.converged should therefore be checked.