Creating work arrays

Required work arrays

The in-place solver! function can be made almost allocation free via the passing of optional pre-allocated work arrays. This can be advantageous if a large number of different diagonalisations are to be performed and the maximum dimensions across all calculations are known ahead of time. The work arrays have a minimum dimension, but not a maximum one. Thus, they can be allocated once, using the maximum dimensions across all calculations to be performed, and then used in every calculation.

Two work vectors have to be supplied:

Twork::Vector{T}, where T<:AllowedTypes

Rwork::Vector{R}, where R<AllowedFloat

where the types T and R are compatible (see Allowed types)

Generating the work arrays

The required dimensions of the Twork and Rwork vectors are dependent on the size of the matrix being diagonalised, the block size, and the maximum subspace dimension. There exists a function, workarrays, that will return the correctly dimensioned arrays given these values.

Davidson.workarrays — Function

workarrays(T, matdim, blocksize, maxvec)

Constructs the Twork and Rwork work arrays required to make the in-place solver! function allocation-free.

Arguments

T<:AllowedTypes: Matrix type

matdim::Int64: Dimension of the matrix

blocksize::Int64: Block size

maxvec::Int64: Maximum subspace dimension

Return values

The return value is of the form Twork, Rwork = workarrays(…), where

Twork::Vector{T<:AllowedTypes}

Rwork::Vector{R<:AllowedFloat}, where R is compatible with T