solver

This section details the use of the solver function, which computes a given number of the lowest-lying eigenpairs of a given matrix A in a 'matrix-free' manner. That is, only a function returning matrix-vector products is required as input, as opposed to the matrix itself.

Davidson.solver — Function

solver(f, diag, nroots, matdim; tol=1e-4, blocksize=nroots+5,
       maxvec=4*blocksize, niter=100, verbose=false)

Arguments

f::Function: In-place matrix-vector multiplication function

diag::Matrix{T}: Diagonal of the matrix whose eigenpairs are sought, where T<:AllowedTypes

nroots::Int64: Number of eigenpairs to compute

matdim::Int64: Dimension of the matrix

Optional keyword arguments

tol::Float64: Residual norm convergence threshold

blocksize::Int64: Block size

maxvec::Int64: Maximum subspace dimension

niter::Int64: Maximum number of iterations

verbose::Bool: Verbose output flag. If true, then a summary is printed at the end of each iteration

Return values

The return value is of the form vectors, values = solver(…), where

vectors::Matrix{T}: Matrix of eigenvectors

values::Vector{R}: Vector of eigenvalues

A note on the matrix-vector multiplication function

Let the matrix whose eigenpairs are sought be denoted by A. The in-place Function f recieves takes as arguments an AbstractMatrix{T} of vectors v and an AbstractMatrix{T} of matrix-vector products Av that is to be computed, where T<:AllowedTypes. The required form of this function is detailed in Matrix-vector multiplication function section.